requirements.txt
PyQt5==5.9.2
pyserial==3.4
sip==4.19.8
